feat: CRUD for Membership

+++ Core/src/main/java/id_authentication/service/MembershipService.java
@@ -4,6 +4,7 @@ import id_authentication.domain.Member;
 import id_authentication.domain.Membership;
 import id_authentication.dto.request.MembershipRequestDto;
 import id_authentication.dto.response.MembershipResponseDto;
+import id_authentication.exceptions.ResourceNotFoundException;
 import id_authentication.repositories.MembershipRepository;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
@@ -11,44 +12,65 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
-public class MembershipService implements IMembershipService{
+public class MembershipService implements IMembershipService {
 
     private final ModelMapper modelMapper;
     private final MembershipRepository membershipRepository;
+
     @Override
     public MembershipResponseDto save(MembershipRequestDto membershipRequestDto) {
         Membership membership = modelMapper.map(membershipRequestDto, Membership.class);
-        membershipRepository.save(membership);
-        return modelMapper.map(membershipRepository.save(membership), MembershipResponseDto.class);
+        Membership savedMembership = membershipRepository.save(membership);
+        membershipRepository.updateMemberId(savedMembership.getId(), membershipRequestDto.getMemberId());
+        return modelMapper.map(savedMembership, MembershipResponseDto.class);
     }
 
-    public MembershipResponseDto getMembership(long id){
-        Membership membership = membershipRepository.findById(id).get();
+    @Override
+    public MembershipResponseDto getMembership(long id) {
+        Membership membership = membershipRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Membership not found with id " + id));
         return modelMapper.map(membership, MembershipResponseDto.class);
     }
 
     public List<Membership> getAllMemberships() {
         List<Membership> membershipsList = new ArrayList<Membership>();
-        membershipRepository.findAll().forEach(membership->membershipsList.add(membership));
+        membershipRepository.findAll().forEach(membership -> membershipsList.add(membership));
         return membershipsList;
     }
 
     @Override
     public MembershipResponseDto updateMembership(long id, MembershipRequestDto membershipRequestDto) {
-        return null;
+        Optional<Membership> membershipOptional = membershipRepository.findById(id);
+        if (membershipOptional.isPresent()) {
+            Membership membership = membershipOptional.get();
+            membership.setStartDate(membershipRequestDto.getStartDate());
+            membership.setEndDate(membershipRequestDto.getEndDate());
+            membership.setMembershipNumber(membershipRequestDto.getMembershipNumber());
+            membership.setType(membershipRequestDto.getType());
+            return modelMapper.map(membershipRepository.save(membership), MembershipResponseDto.class);
+        } else {
+            throw new ResourceNotFoundException("Membership does not exist!!");
+        }
     }
 
 
-    public void deleteMembership(long id) {
-        membershipRepository.deleteById(id);
+    public String deleteMembership(long id) {
+        Optional<Membership> membershipOptional = membershipRepository.findById(id);
+        if (membershipOptional.isPresent()) {
+            membershipRepository.deleteById(id);
+            return "Membership deleted";
+        } else {
+            throw new ResourceNotFoundException("Couldn't find the membership with id: " + id);
+        }
     }
 
     @Override
-    public List<MembershipResponseDto> findAllByMembershipNumber(String memberNumber) {
-
+    public List<MembershipResponseDto> findAllByMemberId(String memberId) {
+        //TODO
         return null;
     }
